begin redirect branch

+++ user/TFT/UIfloatingWindow.h
@@ -4,7 +4,7 @@
 #include "interface.h"
 #include "UIobject.h"
 
-extern UIobject *flotingUI(int centerx, int centery, int width, int height, u16 color, u16 backgroudColor, int priority);
+extern UIobject *flotingUI(short centerx, short centery, short width, short height, u16 color, u16 backgroudColor, int priority);
 extern void floatingMessage(char *message);
 extern void resetFloatingWindowCounter();
 extern void updateMessage(char *message);
